+++ infra/cdk_src/smstudio/cr_lambda_common.py
@@ -29,8 +29,9 @@ class SMCustomResourceHelperLayer(PythonLayerVersion):
         bundling: BundlingOptions | Dict[str, Any] | None = None,
         compatible_architectures: Sequence[Architecture] | None = None,
         compatible_runtimes: Sequence[Runtime] | None = None,
-        description: str
-        | None = ("Helper functions & classes for SageMaker CloudFormation custom resources"),
+        description: str | None = (
+            "Helper functions & classes for SageMaker CloudFormation custom resources"
+        ),
         layer_version_name: str | None = None,
         license: str | None = None,
         removal_policy: RemovalPolicy | None = None,
